Moves this to the more standard Django location
